feat: add service layer for courses and instructors
Introduces CourseService and InstructorService interfaces and their implementations, providing business logic for course and instructor management. Updates DTOs and entity to use Long for instructor IDs, and fixes CourseMapper to use UpdateCourseDTO for entity updates. This refactor enables separation of concerns and prepares for more robust service operations.

+++ src/main/java/com/hibernate/coursemanagement/dto/CreateCourseDTO.java
@@ -40,5 +40,5 @@ public class CreateCourseDTO {
     private Boolean isActive;
 
     @NotNull(message = "Instructor ID is required")
-    private Integer instructorId;
+    private Long instructorId;
 }
