+++ src/main/java/com/hibernate/coursemanagement/dto/UpdateCourseDTO.java
@@ -40,5 +40,5 @@ public class UpdateCourseDTO {
     private Boolean isActive;
 
     @NotNull(message = "Instructor ID is required")
-    private Integer instructorId;
+    private Long instructorId;
 }
